requirements.txt
pyspark==3.5.0
psycopg2-binary==2.9.9
pandas==2.2.0
numpy==1.26.0
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.23 